requirement.txt
requests==2.28.1
google-auth==2.18.0
google-auth-oauthlib==0.5.2
google-api-python-client==2.80.0


